+++ CHANGELOG.md
@@ -1,6 +1,14 @@
-# Version 0.5.2
+# Version 0.6.0
 
+- Polygon inference
+- Unittests to Polygon inference;
 - Bug fixes warmup callback (invalid signature on method);
+- FrameFieldResultCallback renamed to FrameFieldOverlayedResultCallback;
+- New implementation of FrameFieldResultCallback;
+- Invalid mask handling (frame field training mask with only polygon mask and empty vertex and boundary masks);
+- Added multiple schedulers option;
+- Added IoU 10, 25, 50, 75 and 90;
+- Added GPU augmentation using kornia;
 
 # Version 0.5.1